Refinamentos c(:

+++ src/SGP.Shared/AppSettings/AuthConfig.cs
@@ -2,9 +2,9 @@ namespace SGP.Shared.AppSettings;
 
 public sealed class AuthConfig
 {
-    public short MaximumAttempts { get; private set; }
-    public short SecondsBlocked { get; private set; }
+    public int MaximumAttempts { get; private set; }
+    public int SecondsBlocked { get; private set; }
 
-    public static AuthConfig Create(short maximumAttempts, short secondsBlocked)
+    public static AuthConfig Create(int maximumAttempts, int secondsBlocked)
         => new() { MaximumAttempts = maximumAttempts, SecondsBlocked = secondsBlocked };
 }
